+++ component/include/pl_lock_guard.h
@@ -10,20 +10,23 @@ namespace PL {
 /// @brief RAII-style lock guard class
 class LockGuard {
 public:
-  /// @brief Locks the object with infinite timeout
+  /// @brief Locks the object
   /// @param lockable lockable object
-  LockGuard(Lockable& lockable);
+  /// @param timeout timeout in FreeRTOS ticks
+  LockGuard(Lockable& lockable, TickType_t timeout = portMAX_DELAY);
 
-  /// @brief Locks multiple objects with infinite timeout preventing deadlock
+  /// @brief Locks multiple objects preventing deadlock
   /// @param lockable1 lockable object 1
   /// @param lockable2 lockable object 2
-  LockGuard(Lockable& lockable1, Lockable& lockable2);
+  /// @param timeout total timeout in FreeRTOS ticks
+  LockGuard(Lockable& lockable1, Lockable& lockable2, TickType_t timeout = portMAX_DELAY);
 
-  /// @brief Locks multiple objects with infinite timeout preventing deadlock
+  /// @brief Locks multiple objects preventing deadlock
   /// @param lockable1 lockable object 1
   /// @param lockable2 lockable object 2
   /// @param lockable3 lockable object 3
-  LockGuard(Lockable& lockable1, Lockable& lockable2, Lockable& lockable3);
+  /// @param timeout total timeout in FreeRTOS ticks
+  LockGuard(Lockable& lockable1, Lockable& lockable2, Lockable& lockable3, TickType_t timeout = portMAX_DELAY);
 
   ~LockGuard();
   LockGuard(const LockGuard&) = delete;